refactor: distanceMeters hook 분리
