+++ src/lib/hooks/distanceMeters.ts
@@ -1,0 +1,19 @@
+type LatLng = { lat: number; lng: number };
+
+export function distanceMeters(a: LatLng, b: LatLng) {
+  const R = 6371e3;
+  const toRad = (v: number) => (v * Math.PI) / 180;
+
+  const dLat = toRad(b.lat - a.lat);
+  const dLng = toRad(b.lng - a.lng);
+
+  const lat1 = toRad(a.lat);
+  const lat2 = toRad(b.lat);
+
+  const x =
+    Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+    Math.cos(lat1) * Math.cos(lat2) * Math.sin(dLng / 2) * Math.sin(dLng / 2);
+
+  const c = 2 * Math.atan2(Math.sqrt(x), Math.sqrt(1 - x));
+  return R * c;
+}
